requirements.txt
asgiref==3.8.1
Django==4.2.16
djangorestframework==3.15.2
mysqlclient==2.2.4
numpy==2.0.2
pandas==2.2.2
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.16.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
